2019 IntComputer refactoring - swapped IntArray out for LongArray to handle large numbers

+++ src/y2019/day7/day7.kt
@@ -10,16 +10,16 @@ fun main() {
     println(permutations(5 .. 9).map { part2(it, program) }.max())
 }
 
-private fun part1(config: List<Int>, program: String): Int {
-    val amplifiers = config.map { IntComputer(program).addInput(it) }
-    return amplifiers.fold(0) { acc, curr ->
+private fun part1(config: List<Int>, program: String): Long {
+    val amplifiers = config.map { IntComputer(program).addInput(it.toLong()) }
+    return amplifiers.fold(0L) { acc, curr ->
         curr.addInput(acc).run().outputs.first()
     }
 }
 
-private fun part2(config: List<Int>, program: String): Int {
-    val amplifiers = config.map { IntComputer(program).addInput(it) }
-    var output = listOf(0)
+private fun part2(config: List<Int>, program: String): Long {
+    val amplifiers = config.map { IntComputer(program).addInput(it.toLong()) }
+    var output = listOf(0L)
     while (!amplifiers.last().isDone) {
         output = amplifiers.fold(output) { acc, curr ->
             curr.addInput(acc).run().outputs.also { curr.outputs = mutableListOf() }
